+++ src/Application/Product/ProductApplication.php
@@ -8,6 +8,7 @@ use Thinktomorrow\Trader\Application\Product\UpdateProduct\UpdateProductOptions;
 use Thinktomorrow\Trader\Application\Product\UpdateProduct\UpdateProductPersonalisations;
 use Thinktomorrow\Trader\Application\Product\UpdateProduct\UpdateProductTaxa;
 use Thinktomorrow\Trader\Application\Product\UpdateVariant\UpdateVariantOptionValues;
+use Thinktomorrow\Trader\Application\Product\VariantProperties\CleanupRemovedVariantProperties;
 use Thinktomorrow\Trader\Domain\Common\Event\EventDispatcher;
 use Thinktomorrow\Trader\Domain\Model\Product\Events\ProductDeleted;
 use Thinktomorrow\Trader\Domain\Model\Product\Option\Option;
@@ -27,13 +28,15 @@ class ProductApplication
     private EventDispatcher $eventDispatcher;
     private ProductRepository $productRepository;
     private VariantRepository $variantRepository;
+    private CleanupRemovedVariantProperties $cleanupRemovedVariantProperties;
 
-    public function __construct(TraderConfig $traderConfig, EventDispatcher $eventDispatcher, ProductRepository $productRepository, VariantRepository $variantRepository)
+    public function __construct(TraderConfig $traderConfig, EventDispatcher $eventDispatcher, ProductRepository $productRepository, VariantRepository $variantRepository, CleanupRemovedVariantProperties $cleanupRemovedVariantProperties)
     {
         $this->traderConfig = $traderConfig;
         $this->eventDispatcher = $eventDispatcher;
         $this->productRepository = $productRepository;
         $this->variantRepository = $variantRepository;
+        $this->cleanupRemovedVariantProperties = $cleanupRemovedVariantProperties;
     }
 
     public function createProduct(CreateProduct $createProduct): ProductId
@@ -95,7 +98,11 @@ class ProductApplication
     {
         $product = $this->productRepository->find($updateProductTaxa->getProductId());
 
+        $oldProductTaxa = $product->getProductTaxa();
+
         $product->updateProductTaxa($updateProductTaxa->getProductTaxa());
+        
+        $this->cleanupRemovedVariantProperties->handle($product, $oldProductTaxa);
 
         $this->productRepository->save($product);