+++ src/Domain/Model/Product/ProductTaxa/VariantProperty.php
@@ -1,0 +1,11 @@
+<?php
+declare(strict_types=1);
+
+namespace Thinktomorrow\Trader\Domain\Model\Product\ProductTaxa;
+
+/**
+ * Available variant property for a product
+ */
+class VariantProperty extends ProductTaxon
+{
+}