+++ tests/Infrastructure/Repositories/GridRepositoryTest.php
@@ -8,6 +8,7 @@ use Money\Money;
 use Tests\Infrastructure\TestCase;
 use Thinktomorrow\Trader\Application\Product\Grid\GridItem;
 use Thinktomorrow\Trader\Application\Product\ProductApplication;
+use Thinktomorrow\Trader\Application\Product\VariantProperties\CleanupRemovedVariantProperties;
 use Thinktomorrow\Trader\Application\Taxon\TaxonApplication;
 use Thinktomorrow\Trader\Application\Taxonomy\TaxonomyApplication;
 use Thinktomorrow\Trader\Infrastructure\Laravel\Models\DefaultGridItem;
@@ -154,7 +155,7 @@ class GridRepositoryTest extends TestCase
 
         $this->assertCount(3, $gridItems);
 
-        $titles = $gridItems->map(fn ($gridItem) => $gridItem->getTitle());
+        $titles = $gridItems->map(fn($gridItem) => $gridItem->getTitle());
 
         $expected = $titles->toArray();
         natcasesort($expected);
@@ -168,7 +169,7 @@ class GridRepositoryTest extends TestCase
 
         $this->assertCount(3, $gridItems);
 
-        $titles = $gridItems->map(fn ($gridItem) => $gridItem->getTitle());
+        $titles = $gridItems->map(fn($gridItem) => $gridItem->getTitle());
 
         $expected = $titles->toArray();
         natcasesort($expected);
@@ -192,6 +193,7 @@ class GridRepositoryTest extends TestCase
                 new EventDispatcherSpy(),
                 new MysqlProductRepository(new MysqlVariantRepository(new TestContainer())),
                 new MysqlVariantRepository(new TestContainer()),
+                new CleanupRemovedVariantProperties()
             ),
             new MysqlProductRepository(new MysqlVariantRepository(new TestContainer()))
         );