+++ tests/Unit/Model/Product/ProductTaxonTest.php
@@ -8,6 +8,7 @@ use Thinktomorrow\Trader\Domain\Model\Product\Events\ProductCreated;
 use Thinktomorrow\Trader\Domain\Model\Product\Events\ProductTaxaUpdated;
 use Thinktomorrow\Trader\Domain\Model\Product\ProductId;
 use Thinktomorrow\Trader\Domain\Model\Product\ProductTaxa\ProductTaxon;
+use Thinktomorrow\Trader\Domain\Model\Product\ProductTaxa\VariantProperty;
 use Thinktomorrow\Trader\Domain\Model\Taxon\TaxonId;
 
 class ProductTaxonTest extends TestCase
@@ -38,5 +39,16 @@ class ProductTaxonTest extends TestCase
         ]);
 
         $this->assertCount(1, $product->getChildEntities()[ProductTaxon::class]);
+    }
+
+    public function test_it_can_have_variant_properties(): void
+    {
+        $product = $this->createProduct();
+
+        $product->updateProductTaxa([
+            $variantProperty = VariantProperty::create($product->productId, TaxonId::fromString('ppp')),
+        ]);
+
+        $this->assertEquals([$variantProperty], $product->getVariantProperties());
     }
 }