+++ tests/Unit/Model/Product/VariantTaxonTest.php
@@ -11,12 +11,13 @@ use Thinktomorrow\Trader\Domain\Model\Product\Variant\Variant;
 use Thinktomorrow\Trader\Domain\Model\Product\Variant\VariantId;
 use Thinktomorrow\Trader\Domain\Model\Product\Variant\VariantSalePrice;
 use Thinktomorrow\Trader\Domain\Model\Product\Variant\VariantUnitPrice;
+use Thinktomorrow\Trader\Domain\Model\Product\VariantTaxa\VariantProperty;
 use Thinktomorrow\Trader\Domain\Model\Product\VariantTaxa\VariantTaxon;
 use Thinktomorrow\Trader\Domain\Model\Taxon\TaxonId;
 
-class VariantPropertyTest extends TestCase
+class VariantTaxonTest extends TestCase
 {
-    public function test_a_variant_can_have_variant_properties()
+    public function test_a_variant_can_have_variant_taxa()
     {
         $variant = Variant::create(
             $productId = ProductId::fromString('xxx'),
@@ -30,9 +31,37 @@ class VariantPropertyTest extends TestCase
             'sku',
         );
 
-        $variantProperty = VariantTaxon::create(
+        $variantTaxon = VariantTaxon::create(
             $variantId,
-            $taxonId = TaxonId::fromString('bbb')
+            TaxonId::fromString('bbb')
+        );
+
+        $variant->updateVariantTaxa([
+            $variantTaxon,
+        ]);
+
+        $this->assertEquals([
+            $variantTaxon,
+        ], $variant->getVariantTaxa());
+    }
+
+    public function test_it_can_have_variant_properties(): void
+    {
+        $variant = Variant::create(
+            $productId = ProductId::fromString('xxx'),
+            $variantId = VariantId::fromString('yyy'),
+            $productUnitPrice = VariantUnitPrice::fromMoney(
+                Money::EUR(10),
+                VatPercentage::fromString('20'),
+                false
+            ),
+            $productSalePrice = VariantSalePrice::fromMoney(Money::EUR(8), VatPercentage::fromString('20'), false),
+            'sku',
+        );
+
+        $variantProperty = VariantProperty::create(
+            $variantId,
+            TaxonId::fromString('bbb')
         );
 
         $variant->updateVariantTaxa([
@@ -41,6 +70,6 @@ class VariantPropertyTest extends TestCase
 
         $this->assertEquals([
             $variantProperty,
-        ], $variant->getVariantTaxa());
+        ], $variant->getVariantProperties());
     }
 }